feat: make loan manager generic to AMM
This commit makes the LoanManager class more generic to AMM
implementation. Instead of directly taking in a UniswapV3 Pool object,
it now only relies on a thin interface for AMMs. This interface should
work out-of-the-box with both UniV3 and UniV2 and can be implemented
by our own concrete classes for other AMM types