fix: min collateral for borrow
this commit fixes the calculation for minimum collateral on borrow.
Instead of always using the yTranche (i.e. B tranche in an ABZ bond), it
uses the tranche with the highest discount to calculate an estimation.
This should result in the estimation always being more than enough to
support the desired loan amount

It also fixes the getRequiredDeposit calculation to properly use dcr for
tranche ratio expansion rather than cdr